MLService/requirements.txt
fastapi
uvicorn[standard]
rasterio
numpy
pillow
shapely
fiona
geopandas
boto3
pydantic
python-multipart
